roles filter refactor
if the block is in a course, it will consider users with any
of the selected roles assigned in the course or in any sub contexts.

if the block is in a category, it will consider users with any
of the selected roles assigned in the category or in any sub contexts.

if the block is in the dashboard, it will consider users with any
of the selected roles assigned in any context.

if the block is in the system context, it will consider users with any
of the selected roles assigned in any context.